[skip ci] Rename awkward1.behavior -> awkward1.behaviors.
