+++ awkward1/operations/convert.py
@@ -77,9 +77,9 @@ def tonumpy(array):
 
     elif awkward1.operations.describe.parameters(array).get("__array__") == "string":
         if awkward1.operations.describe.parameters(array.content).get("encoding") is None:
-            return numpy.array([awkward1.behavior.string.CharBehavior(array[i]).__bytes__() for i in range(len(array))])
+            return numpy.array([awkward1.behaviors.string.CharBehavior(array[i]).__bytes__() for i in range(len(array))])
         else:
-            return numpy.array([awkward1.behavior.string.CharBehavior(array[i]).__str__() for i in range(len(array))])
+            return numpy.array([awkward1.behaviors.string.CharBehavior(array[i]).__str__() for i in range(len(array))])
 
     elif isinstance(array, awkward1._util.unknowntypes):
         return numpy.array([])
@@ -150,7 +150,7 @@ def tolist(array):
     elif isinstance(array, numpy.ndarray):
         return array.tolist()
 
-    elif isinstance(array, awkward1.behavior.string.CharBehavior):
+    elif isinstance(array, awkward1.behaviors.string.CharBehavior):
         if array.layout.parameters.get("encoding") is None:
             return array.__bytes__()
         else:
@@ -158,9 +158,9 @@ def tolist(array):
 
     elif awkward1.operations.describe.parameters(array).get("__array__") == "char":
         if awkward1.operations.describe.parameters(array).get("encoding") is None:
-            return awkward1.behavior.string.CharBehavior(array).__bytes__()
+            return awkward1.behaviors.string.CharBehavior(array).__bytes__()
         else:
-            return awkward1.behavior.string.CharBehavior(array).__str__()
+            return awkward1.behaviors.string.CharBehavior(array).__str__()
 
     elif isinstance(array, awkward1.highlevel.Array):
         return [tolist(x) for x in array]
